pyproject.toml
[tool.poetry]
name = "django-rest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.8"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
celery = "^5.4.0"
eventlet = "^0.38.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
